backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.2.0
pyjwt==2.8.0
python-dotenv==1.0.1
gunicorn==21.2.0
passlib[bcrypt]==1.7.4
